rpi_requirements.txt
actionlib==1.13.2
angles==1.9.13
bondpy==1.8.6
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
casadi==3.5.5
catkin==0.8.10
catkin-pkg==0.4.24
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cv-bridge==1.16.0
cycler==0.11.0
defusedxml==0.7.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.6.0
do-mpc==4.2.5
docutils==0.18.1
empy==3.3.4
fonttools==4.28.2
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
image-geometry==1.16.0
interactive-markers==1.12.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.3.2
laser_geometry==1.6.7
matplotlib==3.5.0
message-filters==1.15.14
numpy==1.19.5
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-qt-binding==0.4.4
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
quadprog==0.1.11
resource_retriever==1.12.7
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.3.0
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-moveit==0.5.10
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
ruckig==0.5.0
scipy==1.7.3
sensor-msgs==1.13.1
setuptools-scm==6.3.2
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
tomli==1.2.2
topic-tools==1.15.14
xacro==1.13.3
